working admin panel

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -2,8 +2,8 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Models\Category;
 use App\Http\Controllers\Controller;
+use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
@@ -11,8 +11,6 @@ class CategoryController extends Controller
 {
     /**
      * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
      */
     public function index()
     {
@@ -22,8 +20,6 @@ class CategoryController extends Controller
 
     /**
      * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
      */
     public function create()
     {
@@ -32,24 +28,21 @@ class CategoryController extends Controller
 
     /**
      * Store a newly created resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            $params['image'] = $request->file('image')->store('categories');
+        }
+
         Category::create($params);
         return redirect()->route('categories.index');
     }
 
     /**
      * Display the specified resource.
-     *
-     * @param \App\Category $category
-     * @return \Illuminate\Http\Response
      */
     public function show(Category $category)
     {
@@ -58,9 +51,6 @@ class CategoryController extends Controller
 
     /**
      * Show the form for editing the specified resource.
-     *
-     * @param \App\Category $category
-     * @return \Illuminate\Http\Response
      */
     public function edit(Category $category)
     {
@@ -69,26 +59,22 @@ class CategoryController extends Controller
 
     /**
      * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param \App\Category $category
-     * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Category $category)
     {
-        Storage::delete($category->image);
-        $path = $request->file('image')->store('categories');
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+            $params['image'] = $request->file('image')->store('categories');
+        }
+
         $category->update($params);
         return redirect()->route('categories.index');
     }
 
     /**
      * Remove the specified resource from storage.
-     *
-     * @param \App\Category $category
-     * @return \Illuminate\Http\Response
      */
     public function destroy(Category $category)
     {
